V3.1.0- Guide post fixed & File upload Added

+++ src/tour-guide-post/tour-guide-post.service.ts
@@ -29,6 +29,7 @@ export class TourGuidePostService {
     // Create the TourPost and link it to the TourGuide (user)
     const tourPost = this.tourPostRepository.create({
         ...createTourPost, // Data from body and checked by DTO
+        //V3.1.0- Fixing the createdBy field to be the TourGuide id instead of the user id
         createdBy: tourGuide, // Linking the post to the authenticated user (TourGuide)
       });
 
@@ -46,7 +47,7 @@ export class TourGuidePostService {
     }
 
     return this.tourPostRepository.find({
-        where: {createdBy: { id: guideId-1 }, // Filtering by the TourGuide's userId | Update the query was returning id = 2 for token of id=1. I dont know why. Deducing 1 from the id to get the correct id of the userid output resulted in solving the issue!
+        where: {createdBy: { id: guideId }, // Filtering by the TourGuide's userId | Update the query was returning id = 2 for token of id=1. I dont know why. Deducing 1 from the id to get the correct id of the userid output resulted in solving the issue!
         },
         relations: ['createdBy'], // Ensuring the 'createdBy' field inclusion
       });
@@ -54,12 +55,26 @@ export class TourGuidePostService {
   
 
   //v2.0.0- Update Tour Guide Post
-  async updateTourGuide(guidePostId: number,updateGuide:UpdateGuidePostDto,): Promise<GUIDE_POST> {
+  async updateTourGuide(guideID: number, guidePostId: number, updateGuide: UpdateGuidePostDto,): Promise<GUIDE_POST> {
     
-    const guidePost = await this.tourPostRepository.findOneBy({id: guidePostId,});
+    // const guidePost = await this.tourPostRepository.findOneBy({id: guidePostId,}); //This does not work
+    //Why? Because findOneBy is not a function of Repository. It should be findOne instead. Plus because createdBy is a relation.
+    
+    //v3.1.0- Fixing the update method to check if the post was created by the authenticated user.
+    const guidePost = await this.tourPostRepository.findOne({
+        where: { id: guidePostId },
+        relations: ['createdBy'], // Ensure 'createdBy' data is included
+    });
 
     if (!guidePost) {
       throw new NotFoundException(`Tour Guide Post not found`);
+    }
+    console.log("Guide Post by who:", guidePost.createdBy.id); // Debugging
+    console.log("Guide update action ID:", guideID); // Debugging
+
+    //V3.1.0-(Authorized update check) Checking if the post was created by the authenticated user
+    if(guidePost.createdBy.id !== guideID){
+      throw new NotFoundException(`You are not authorized to update this post`);
     }
 
     Object.assign(guidePost, updateGuide); // Updating fields with the new values from the DTO
@@ -69,14 +84,21 @@ export class TourGuidePostService {
   }
 
 //   v2.0.0- Delete Tour Guide Post
-  async deleteGuidePost(guidePostId: number) {
-    const guidePost = await this.tourPostRepository.findOneBy({
-      id: guidePostId,
+  async deleteGuidePost(guideID: number, guidePostId: number) {
+    //v3.1.0- Fixing the update method to check if the post was created by the authenticated user
+    const guidePost = await this.tourPostRepository.findOne({
+      where: { id: guidePostId },
+      relations: ['createdBy'], // Ensure 'createdBy' data is included
     });
 
     if (!guidePost) {
       throw new NotFoundException(`Tour Guide Post not found`);
     }
+    //V3.1.0-(Authorized update check) Checking if the post was created by the authenticated user
+    if(guidePost.createdBy.id !== guideID){
+      throw new NotFoundException(`You are not authorized to update this post`);
+    }
+
     return await this.tourPostRepository.remove(guidePost);
   }
 }
